+++ src/background/chrome-shortcuts-page-controller.ts
